Conflicts:
	matlab/deblur.m
	matlab/testBlurDeblur.m

+++ matlab/save_image.m
@@ -9,7 +9,7 @@ if graph == 1 || graph == 2
     end
     colormap(gray);
     title(name);
-    imshow(L/255);
+    imshow(double(L)/255);
     axis off;
     if graph == 1
         saveas(gcf, name, 'png');
